// app/types/profile.ts
export interface User {
  user_id: string;
  full_name: string;
  pronouns?: string;
  consent: boolean;
}

// Uncomment when needed
// export interface Streak {
//   current_streak: number;
//   longest_streak: number;
// }

// export interface Assessment {
//   id: number;
//   date: string;
//   type: string;
//   score: number;
// } 